add importmasterkey, mnemonicinfo rpc commands (#37)
Adds RPC commands for importing the HD Master Key from an
easily memorable seed (BIP39).

The implementation tries to leverage as much of bitcoin-core's
existing infrastructure as possible. One caveat is that it is possible
to import a new master key into an existing and used wallet.
Because of this `importmasterkey` will always backup your wallet.
A ticket has been opened to deal with this case (silently import if
existing wallet is unused/was newly created and fail if is used).