Allow test to fail the generate call
Sometimes it may happen that the generate rpc call fails because the
block being generate contain transactions that are no longer valid
at the moment of the block validation before proposal. To account for
this case in the eventuality that a generate call fails it will try
again few times before giving up.