Remove outdated votes from the mempool (#92)
Fix a bug causing a vote to remain stuck in a mempool. This would happen when a vote after inclusion in the node's mempool becomes invalid (voting an outdated checkpoint). If nobody publishes the vote in a block then at the moment of proposal the node would always generate an invalid block.