Replace travis_wait with a background job (#115)
travis_wait redirects temporary the output to a file and prevents the
current job to fail because of timeouts. At the moment is being used
when calling "make check", this is problematic since if a test fails
travis just kills the build and any output useful to fix the failing
test is gone forever (let alone the test name itself). With this
solution we prevent travis from killing the process at the cost of
mixing the logs between the "keep-alive" background job and the
"make check".