Add Initial Snapshot Download (#149)
* Add Initial Snapshot Download

This commit implements Initial Snapshot Download (ISD) as an alternative
to Initial Block Download (IBD).

To enable ISD node must be started with `-prune=(>0)` and `-isd` flags.
If one of these flags are not set, ISD is disabled.
To disable snapshot creation, start the node with `-createsnapshot=0`. But default it's `1`

If ISD mode is enabled, the node will try to discover the latest snapshot
(max height) from its peers. Then it will download in chunks sequentially
the snapshot and verify that it has the correct checksum. If the snapshot is incorrect,
the node will drop it and re-request a new one. Once the snapshot is considered valid,
the node will ask its peers to give the parent block of the block the snapshot was generated.
When this parent block is received, the node applies snapshot and make the parent block of it the new tip.

Current commit doesn't cover the security part. It assumes that if the snapshot
has the right hash and there is a parent block of it, this snapshot is valid.
Security part will be covered in the future commits.