Add snapshot hash calculation (#174)
This PR introduces `SnapshotHash` class that calculates the hash from UTXOs using `secp256k1_multiset`.

Also, a new struct `UTXO` is introduced which represents a single output, and its serialized data is used to add/subtract outputs from the hash.

Notice: `UTXOSubset` is still used to store a snapshot on disk or transfer it to peers as it has a more compact format. To calculate the hash, `UTXOSubset` must be split by multiple `UTXO` objects. Due to the nature of `secp256k1_multiset` order of operands doesn't matter.