Remove section about finding reviewers
The section explains why it can take a long time to find a reviewer.
We are in a different phase of the project and want to strive for
responding quickly to incoming request. So for now let's assume
that we manage to do that and don't need explanations for lack of
response.