Fix timeouts in commptool framework
In some comptool tests, especially in the `feature_block.py`, the large
re-organization happens during which the node creates many events
that are stored in `g_signals`. When the list is larger than 10,
node doesn't process any new p2p messages until it drains the queue.

To make sure that the node is available for p2p messages, before sending
new blocks check that we didn't exceed the limit of events and if yes,
wait until it the queue is smaller than 10.

Also, before stopping the node, wait until all the events are processed.
In attached logs we can see that stopping the node can take longer than the minute.
```
2018-10-31 15:12:29.780000 TestFramework.comptool (INFO): Running test #96: ./test/functional/feature_block.py line 1268
2018-10-31 15:12:57.369000 TestFramework.comptool (INFO): Running test #97: ./test/functional/feature_block.py line 1273
2018-10-31 15:13:59.681000 TestFramework.comptool (INFO): Running test #98: ./test/functional/feature_block.py line 1275
2018-10-31 15:15:15.962000 TestFramework (INFO): Stopping nodes
2018-10-31 15:16:17.750000 TestFramework (INFO): Cleaning up
2018-10-31 15:16:17.874000 TestFramework (INFO): Tests successful
```