Document maintainers (#188)
Add MAINTAINERS.md as human-readable definition of maintainers. The
file also describes areas which need expert review. We decided to not
add a CODEOWNERS file to automatically assign reviewers as we do want
to operate with the model of shared ownership of the code and don't want
to create silos where only one expert feels responsible.