Allow unlocking wallet for staking only (#239)
When a user decides to encrypt their wallet, they need to provide the
decryption passphrase every time they want to perform a sensitive
operation; the wallet is unlocked for a specified duration and
automatically locked afterwards (the decryption key is removed from
memory).

This obviously makes it difficult to participate in proof-of-stake,
since the user needs to be able to create coinstake transactions to
propose blocks. However, we don't want to force the user to keep the
wallet unlocked for obvious security reasons.

Therefore, following Particl's lead, we introduce a new argument to
`walletpassphrase`, showing whether the wallet is unlocked for staking
only, or for all operations. If the argument is `true`, the node will
participate in block-proposal operations, but all RPC calls which depend
on the wallet being unlocked will return an error.

Now the user may also run `walletpassphrase <...> 0 true` to unlock the
wallet for staking for an unlimited amount of time. Trying to unlock the
wallet for all operations will work as usual (will re-lock immediately).

We also update the `getwalletinfo` RPC call to allow retrieving the
wallet's encryption/lock status, since it's no longer binary.

Signed-off-by: Mihai Ciumeica <mihai@thirdhash.com>