Add support for adding headers to Javascript files
Add license headers to Javascript (`*.js`) and TypeScript (`*.ts`)
files when calling `copyright_heade.py insert`.

This is mainly meant for unit-e-desktop. It adds the GPLv2 header
as this is the license which is used there.

This change also should add support for updating the year range in
copyright headers in Javascript files but this is not tested.

Signed-off-by: Cornelius Schumacher <cornelius@thirdhash.com>