Return all Tx fields in filtertransactions
The `filterMode` argument in WalletTxToJSON controls whether the
transaction fields in `CWalletTx::mapValue` are skipped when serializing
the Tx to JSON. Previously, it was set to `true` when returning the result
of `filtertransactions` (which means that fields like `replaced_by_txid`
were omitted).

Since we now need one of those fields in the desktop wallet, and since the
similar RPC call `listsinceblock` doesn't filter out any fields, it makes
sense to include the fields in `filtertransactions`, as well.

Signed-off-by: Mihai Ciumeica <mihai@thirdhash.com>