Allow estimating RBF fees before committing (#335)
When the user decides to bump the fee on an RBF transaction, and doesn't want to specify an explicit fee, s/he might want to find out what the new fee will be, before committing the transaction. This commit adds a new argument to the bumpfee RPC call, which does exactly that. Fixes #334.

The Qt wallet does this a bit more smoothly: it creates the transaction, asks the user to confirm the new fee, then commits it. The Electron wallet doesn't have access to the internals of the feebumper, so we're forced to create two different transactions (one for estimation, another final), and hope that the estimated fee doesn't change in the 5-10 seconds between them.

Signed-off-by: Mihai Ciumeica <mihai@thirdhash.com>