Reduce global state in snapshot/p2p_processing
This refactoring moves P2P global state to P2PState class.

Motivation is that I am working on extending p2p processing
and need to mock/change time related values in tests.
With the global state it's not so easy to do this but since
the state is moved to the object, it becomes trivial.

Signed-off-by: Kostiantyn Stepaniuk <kostia@thirdhash.com>