Split snapshot message by header and body
This PR introduces a separate two P2P messages
to request/receive the snapshot header:
1. `DISCSNAPSHOT` (discover the peer's best snapshot)
2. `BESTSNAPSHOT` (header of the peer's best snapshot)

There are two reasons for this change:
1. reduce bandwidth. As now we don't need to repeat `block_hash`,
`stake_modifier` and `total_utxo_subsets` in the snapshot chunk response.
We also consider to add `height` as meta-information to the header
and most likely we will have other fields in the future so it's good to separate them.
2. not covered in this PR, but we have a plan to download headers from the peer
which claims has the most recent snapshot. Currently, we download from the random one.
Having a separate P2P message to receive a snapshot header makes a nice logical separation,
will be aligned with the block/header messages.

This PR also changes the download process. Currently, the node asked every peer to
give the snapshot and if the received one is a better one, the node switched to it,
if it's worse, it ignored it. Now node makes the decision based on snapshot headers
which snapshot to download. It also now properly track the timeout of each individual peer
and doesn't prolong it if other peer joined the downloading process.

The fallback is the same as before. If node runs our of peers that can provide
the chosen snapshot, it switches to the second best and so on until it runs
out of peers. If there are no peers that can provide the snapshot, it switches to IBD.

Signed-off-by: Kostiantyn Stepaniuk <kostia@thirdhash.com>