Keep track of node's index and total count in SendMessages
The reason of this change is that we often need to make a decision
after checking the state of all nodes. E.g.: find the best snapshot among nodes.
For such use cases, we use g_connman->ForEachNode() which resolves
to n^2 complexity (outer loop + inner loop) and LOCK on cs_vNodes.

Knowing if we are processing the first or the last block will allow us
to adjust the code to make some initialization when it's the first iteration
of the loop or do cleanup if it's the last one and keep the complexity N

Signed-off-by: Kostiantyn Stepaniuk <kostia@thirdhash.com>