Use DNS to discover nodes with snapshot support
Allow to discover peers that support snapshot via DNS.
If we can't find such peers via DNS, we fallback to previous logic:
discover all the peers that support the full chain either via seeds or other peers.

This should increase the chance of finding the peer that has the snapshot
during the initial fast sync. We would also need to add our nodes to seeds
that support all service flags.

Signed-off-by: Kostiantyn Stepaniuk <kostia@thirdhash.com>