Add configuration of GitHub labels
This adds a YAML configuration of GitHub labels which can be applied
via the API through Terraform and a bit of scripting. The script is
available at https://github.com/cornelius/beautiful-labels.

The configuration includes descriptions for the labels and uses colors
to group categories.

It's the implementation of the label part of ADR-24.

Signed-off-by: Cornelius Schumacher <cornelius@thirdhash.com>