Replace 'HDMasterKeyID' with 'HDSeed'
In preparation for adding hardware wallet support, we rename instances of
'hdmasterkeyid' to 'hdseed', to avoid confusion.

In case of a hardware wallet, we don't know the seed, but we do know the
master key ID. However, right now we can't store it in the HD chain, since
the field means something different.

The rename was already done upstream in bitcoin/bitcoin@131d4450

Signed-off-by: Mihai Ciumeica <mihai@thirdhash.com>