Add nChainWork to snapshot
This is a placeholder for the chain_stake.
Reasoning is described here: https://github.com/dtr-org/unit-e/issues/436

While working on this PR noticed these 2 issues:
1. There are multiple messages that have the same or partially the same schema.
Extracting the common fields into a separate struct would make it easier to
extend the schema of the snapshot overall and make sure that all relevant parts are updated.
2. Most of the functions accept raw fields, e.g.:
```c++
Indexer::Indexer(const uint256 &snapshot_hash, const uint256 &block_hash,
                 const uint256 &stake_modifier, const uint256 &chain_work,
                 uint32_t step, uint32_t steps_per_file)
```
It's easy to pass arguments in the wrong order and introduce a bug.
Would be better to pass objects that already hold these values,
e.g. `CBlockIndex` instead of `block_hash`, `stake_modifier` and `chain_work`.

I intend to work on the refactoring and tackling the above issues in the follow-up PR.

Signed-off-by: Kostiantyn Stepaniuk <kostia@thirdhash.com>