Don't reverse bytes in uint256.GetHex() functions
Since internally the protocol uses little endian
and most of the data are printed without reversing the order of bytes
it makes sense to also not to reverse it for uint256. This will
make debugging easier as you can dump bytes in any language
and simply grep logs using that string.

The change affects only slightly RPC and logs.
We also need to test how block explorer will behave.

I fixed ~half of unit tests and feature_block.py just to give an idea
what changes are needed.

Signed-off-by: Kostiantyn Stepaniuk <kostia@thirdhash.com>