Test that node detects invalid snapshots
This PR checks that the node detects that its downloaded snapshot is invalid,
deletes it and fallbacks to the second best.

This PR also makes adjustment to the node that it waits until all the peers
replied with their best snapshot or timed out and only then starts downloading the
best snapshot. New test caught the issue that node started downloading the second
best just because it didn't receive fast enough the reply from all the peers.

Signed-off-by: Kostiantyn Stepaniuk <kostia@thirdhash.com>