Fix race condition in snapshot creation
Detected a race-condition which happens when one thread is reading the snapshot
and another one creates a new one or deletes an old one.

This PR makes the following changes:
1. Only try to read the snapshot from disk that is registered in SnapshotIndex.
In master, RPC commands always tried to read the snapshot from disk.
If snapshot is missing, we receive nullptr which is OK.
However, if it's a new snapshot that is being created at this moment, we might not
able to read it (unserialize) as the file might not have all the bytes and this crashes the node.
2. Introduce cs_snapshot which must be held during the lifetime of opened snapshot::Index.
The same is true when we delete the snapshot, lock on cs_snapshot must be held. It prevents the crash
when node serves the snapshot to its peers and at this snapshot is being deleted from disk.
3. Small improvement: don't re-create the existing snapshot. It can happen when the re-org goes back and forth.

Added the unit test that reproduces the bug on master

Signed-off-by: Kostiantyn Stepaniuk <kostia@thirdhash.com>