Don't ask for the snapshot if NODE_SNAPSHOT service flag is not set
There is no need to ask the peer for the snapshot if it already claims it doesn't support this service flag.

This PR also tests that once the node downloaded the snapshot it won't discard it
and will keep asking all the peers for the parent block of it. The reason is that
once we decided to download the snapshot, it must be in the finalized epoch so
even if the snapshot is invalid, we still need this block.

Signed-off-by: Kostiantyn Stepaniuk <kostia@thirdhash.com>