LOCK SnapshotIndex during serialization
Adding snapshot hash or removing it from the `SnapshotIndex` is thread-safe
as the `LOCK(m_cs)` is used. However, serialization is not locked
which means the potential data-race can occur on the `std::map` or `std::set`.

There is no problem right now as serialization is happening only in one thread
currently (in unite-snapshot) but as this data structure supposed to be thread-safe,
serialization shouldn't violate this rule.

Signed-off-by: Kostiantyn Stepaniuk <kostia@thirdhash.com>