Unify UIP names in code and comments
The reason of using the format UIP<number> everywhere is that
by a single pattern you can find all the related information about the UIP.

If we search "BIP", it's the same format in the code and comments.

This commit is like an ADR, once it's accepted, we should pay attention
to comments as we will soon have more references to UIPs in the code.

Signed-off-by: Kostiantyn Stepaniuk <kostia@thirdhash.com>