Add --skip-checkout and --hosts options
Adds an option `--skip-checkout` which skips the checkout of the
git dir from which the gitian-build.py code and the gitian descriptors
are taken. This allows to directly build from a modified version
without having to create tentative commits.

Adds an option `--hosts` to provide a list of hosts for which to build.
This allows to selectively only build certain architectures and not
always build for all hosts. This provides more granular control than
just the `--os` option which reduces build times if you are not interested
in all platforms.

Signed-off-by: Cornelius Schumacher <cornelius@thirdhash.com>