Implement Ledger Nano hardware wallet support (#507)
This is the last PR of the series, and it deals with the Ledger Nano S specifically. We implement the `APDU` class, which wraps the hardware wallet command protocol units, extend the `USBDevice` class to add Ledger support, and use [libhidapi](https://github.com/signal11/hidapi/) to send the commands to the wallet and parse the result.

To build the project, you will need to install the `libhidapi-dev` package. By default, hardware wallet support is now disabled, so you will also need to run `./configure --enable-usbdevice`.

Signed-off-by: Mihai Ciumeica <mihai@thirdhash.com>