Remove createsnapshot RPC command
This command was created to simulate snapshot creation when we didn't have
a background job for that. Now all the tests moved to automated snapshot creation
and this one is not used any more.

This also resolves https://github.com/dtr-org/unit-e/issues/560 as
if we don't have test, we don't need to fix it :)

Signed-off-by: Kostiantyn Stepaniuk <kostia@thirdhash.com>