Don't delay deposits because of re-orgs
There is no need to delay deposit processing until a particular height is reached.
It can be included at any height if its input is spendable.

Also, while working on some tests, noticed that it would be convenient
when two validators using the same mnemonics, spend to the same address
can create the same deposit TX ID. Since this nLockTime is randomized, it's a bit trickier
to emulate this state.

Signed-off-by: Kostiantyn Stepaniuk <kostia@thirdhash.com>