Add --skip-checkout, --hosts, and --work-dir options (#551)
Adds an option `--skip-checkout` which skips the checkout of the
git dir from which the gitian-build.py code and the gitian descriptors
are taken. This allows to directly build from a modified version
without having to create tentative commits.

Adds an option `--hosts` to provide a list of hosts for which to build.
This allows to selectively only build certain architectures and not
always build for all hosts. This provides more granular control than
just the `--os` option which reduces build times if you are not interested
in all platforms.

Also adds an option for setting the working directory. Instead of building
in the current working directory, the working directory can be passed in
with the `--work-dir` option. The working directory holds the build data
including a cache of compiled dependencies so reusing an existing
directory is preferable. The old behavior of the script can be achieved by
using `--work-dir .`.

Signed-off-by: Cornelius Schumacher <cornelius@thirdhash.com>