Use BlockValidator::ContextualCheckBlockHeader
- Removes one PoW check in ::ContextualCheckBlockHeader
- Remove bad-diffbits check from feature-block tests

This uses `BlockValidator::ContextualCheckBlockHeader` as a drop-in
replacement in `ContextualCheckBlockHeader`. Thanks to the tests which
were added by @Gnappuraz in #579 the very same tests will also check the new
contextual check block.

This does remove one PoW check. All other PoW checks are still there,
especially in `CheckBlockHeader`.

The new `BlockValidator::ContextualCheckBlockHeader` does invoke `BlockValidator::CheckBlockHeader` if the validation context passed in `ValidationInfo` does not know that it succeeded already. The new `BlockValidator::CheckBlockHeader` does not do a `PoW` check but it checks that the timestamp is well-formed. This check is bypassed and there's UNIT-E TODO slapped on to it.

Disabling the check for `bad-diffbits`. This is the check which was actually removed, it is in contextual check block to check the difficulty against the calulated one. This works completely different in the PoS setting as it affects the kernel protocol.

Signed-off-by: Julian Fleischer <julian@thirdhash.com>