Fix tests that don't work with the finalization
These fixes are extracted from the https://github.com/frolosofsky/unit-e/pull/1/commits PR
and the purpose is to reduce the review burden on https://github.com/dtr-org/unit-e/pull/525

While working on the fork-choice rules, noticed that there are tests which don't work with the finalization,
e.g., try to re-org before finalization.

The workaround is to increase epoch length to 99999 to ensure that finalization doesn't happen.

These tests are:
1. wallet_import_rescan.py
2. rpc_blockchain.py
3. rpc_getchaintips.py
4. mempool_reorg.py

Signed-off-by: Kostiantyn Stepaniuk <kostia@thirdhash.com>