Allow ignoring remotely staked coins when spending UTE
This commit adds a parameter to the coin control structure, which tells
the daemon to skip remotely staked outputs when selecting the coins
available for spending. (The default behavior is to spend them together
with the others).

It would be a good idea to de-prioritize those coins even if no coin
control parameters are given.

Signed-off-by: Mihai Ciumeica <mihai@thirdhash.com>