Fix epoch calculations
We have several bugs in epoch calculations which are addressed in this PR:
1. target epoch of the vote must be the epoch of the last checkpoint, not current epoch
2. target hash of the vote must be the checkpoint of the target epoch and validated accordingly
3. source epoch must be the epoch of the last justification and not the `current epoch - 1`
4. dynasty must be incremented when `current epoch - 3` is finalized, not `current epoch - 2` because: `finalized (0) - justified (1) - votes for justification (2) - current epoch (3)`
5. Fix `IsVoteExpired`.  Must be `vote.m_targetEpoch < state->GetCurrentEpoch() - 1` instead of `vote.m_targetEpoch < state->GetCurrentEpoch()` otherwise, we keep in the mempool maximum 1 vote (the last one).
6. Replace `InstaFinalize` with `InstaJustify` as it's not consistent with how validators behave and as soon as we leave instant finalization mode, we end up in an inconsistent state. This feature should be completely removed https://github.com/dtr-org/unit-e/issues/572

There is a known issue that since we start with the `epoch=0` and it's not finalized, and the epoch type is `uint32` it means that `last_justified_epoch=0` can mean two things:
there are no justifications at all or we justified the first epoch. The right approach would be to switch to `int64` and start counting from `-1` or use `boost::optional`.
For now, we will keep two zero justified epochs (blocks in a range 0..4 and 5..9) as a workaround and we will apply the proper fix once we start working on https://github.com/dtr-org/unit-e/issues/570 (make the genesis part of finalized epoch=0 and the first block is in epoch=1)

Signed-off-by: Kostiantyn Stepaniuk <kostia@thirdhash.com>