Fix floating esperanza_vote.py
Resolves #602

I reworked the esperanza_vote.py test to have defined point
at which finalizers are voting. That should fix the issue of the floating test.

If this approach is approved, I'll apply it to other tests
and eventually delete `generate_sync` function as it doesn't guarantee
that all the messages are processed sequentially in an expected order
and can cause the failure. (e.g., vote arrived from the future)

Signed-off-by: Kostiantyn Stepaniuk <kostia@thirdhash.com>