Synchronize wallet on generateBlocks and move generatetoaddress to rpcwallet.cpp (#691)
In the generateBlock, there is a race between getting eligible coins (generateBlock) and updating the coins (BlockConnected signal). It is possible that the coin used for staking was marked as spent, but the output of the block was not added to the available coins, and we . We need to wait until wallet has updated to the current tip before trying to generate next block.
This is not the case in Bitcoin, as they do not depend on any of the wallet functionalities in the generate* rpc calls.

         if (pwallet)
            pwallet->BlockUntilSyncedToCurrentChain(); 

Moves the generatetoaddress to rpcwallet.cpp, as it needs access to the wallet, and generateBlocks is only called from there now, so moves it as well. This has the benefit of getting rid of dependency on the walle in rpc/mining.cpp (no more #ifdef ENABLE_WALLET).

The wallet_disable.py functional test also has to be adjusted, as we are not able to generate blocks without a wallet.

Signed-off-by: Mateusz Morusiewicz <mateusz@thirdhash.com>