Set the key time to 1 for keys derived from imported seeds
This solves a bug I've encountered while fixing the new coinbase tests:
when importing a new master key mnemonic, the blockchain is rescanned for
transactions that are using the keys derived from this seed.

However, the key time for the derived keys is still set to the current time, so
if ever the wallet has to be rescanned for transactions (e.g. when the node is
relaunched with the `-zapwallettxes`), the blockchain will only be rescanned up
to the moment when the seed was imported, and any previous balance will remain
unaccounted for.

This patch sets the key time to `1` for the first 1000 keys derived after
importing a master key. The user may pass the `brand_new` parameter to the
`importmnemonic` RPC call to avoid full blockchain rescans in the future.

Signed-off-by: Mihai Ciumeica <mihai@thirdhash.com>