Fix IsSpent for coinbase
In case of a reorg, the coinbases were not handled properly.
In bitcoin it does not matter, as the coinbase never spends anything.
Once we start having inputs in coinbases, this check fails when the
depth is 0 (block that the coinbase in question is in is no longer part
of the current chain), as it marks coins as spent by this coinbase,
even though it should not, as the coinbase will not be applied as
the transactions in mempool.

Signed-off-by: Mateusz Morusiewicz <mateusz@thirdhash.com>