Improve IBLT (#694)
This PR introduces several improvements to IBLT implementation. Although those improvements are more actual in context of graphene #566, this PR can be merged both before and after it. Please, don't be afraid of big size of this PR. This is mostly an IBLTParams table moved.

1) Addresses several nits that were suggested after initial implementation was merged (or just missed)
2) Uses CompactSize for IBLT entry `count` field. I have observed, that it is very rare for `count` to be bigger than 253. In fact, largest value I have seen in tests was 353. And fraction of such big numbers was very low. CompactSize allows to reduce entry size from 16 bytes to 13 bytes on average
3) Optimizes Murmur hash computation for IBLT.
Previously, `MurmurHash3` was accepting vector of bytes. This forced us to convert data to vector even though vector was not really needed. This extra allocation happened dozens of times for each IBLT update. Optimized version completely avoids extra allocations and works almost twice as fast for IBLT Insert/Erase operations

Signed-off-by: Aleksandr Mikhailov <aleksandr@thirdhash.com>