Use SFINAE to generalize util::to_string (#706)
`BOOST_CHECK_EQUAL` requires to be able to stringify its arguments for presenting errors. This in turn works for us due to a nifty definition in `test_unite.h`:

```C++
// To extend boost logging with custom types, we have to extend <<(std::ostream, T).
// Extention must be defined in the same namespace as T.
// To extend << for std::vector, we have to put operator<< in namespace std.
// https://www.boost.org/doc/libs/1_69_0/libs/test/doc/html/boost_test/test_output/test_tools_support_for_logging/testing_tool_output_disable.html
namespace std {
template <typename T>
::std::ostream &operator<<(::std::ostream &os, const ::std::vector<T> &v) {
    os << ::util::to_string(v);
    return os;
}
template <typename T>
::std::ostream &operator<<(::std::ostream &os, const ::std::set<T> &v) {
  os << ::util::to_string(v);
  return os;
}
}
```

These in turn rely on `util::to_string`. The existing definitions were not very general and did not apply to all container types, in particular as they rely on a mixture of overloading and template substitution.

I refactored these definitions using SFINAE (`std::enable_if`) into the 5 presented in this pull request. These

1. Turn anything that provides a `ToString` method into a string
2. Turn anything that can be passed to `std::to_string` into a string
3. Turn strings into strings (this is necessary as otherwise (5) kicks in and serializes it like a container)
4. Turns `std::pair` into a string (necessary for turning `std::map` and the like into a string)
5. A general template that applies to anything that provides `begin()` and `end()` iterator accessors.

I provided a test which shows how an `std::map` with nested `std::vector` and `std::set` and a thing that provides `ToString()` is stringified.

The construction with `util::to_string` and `util::stringify` is necessary as the parameters in `std::enable_if` are not in a deducible context, that is the type parameters have to be specified, leading to duplication of code in the call site. The new `util::to_string` fixes the type parameter and specifies it explicitly in order to call `util::stringify`.

As can be seen in `proof_of_stake` tests this leads to some simpler comparing of types. Also this will be useful in log statements, simply slap `util::to_string` on anything. If there's a way to turn it into a string it will be turned into a string.

Signed-off-by: Julian Fleischer <julian@thirdhash.com>

Signed-off-by: Julian Fleischer <julian@thirdhash.com>