Redefine IsCoinbase() using type (#724)
Bitcoin defines the coinbase as just the tx as `(vin.size() == 1 && vin[0].prevout.IsNull())`.
We have transaction types, hence is correct to define the coinbase using that.

Note: I'm momentarily disabling the `tx_valid` test. This breaks using a different type bit and I'll fix it and reenable it once we made all changes to the transaction structure that pos requires. 

Signed-off-by: Matteo Sumberaz <matteo@thirdhash.com>
Signed-off-by: Julian Fleischer <julian@thirdhash.com>