Fix proposer_settings.py
Now that coinbase transactions have inputs, the fee calculation logic
results in the block reward being returned in the 'fee' field (and
the 'amount' field being equal to debit - credit - fee = 0).
Previously, the fee was 0 and the block reward was returned as 'amount'.

It's probably easier to accept the new behavior and fix the tests
accordingly, rather than change the code to restore the old one.

Signed-off-by: Mihai Ciumeica <mihai@thirdhash.com>