Remove Linux packaging files
These are outdated upstream files which were not adapted to unit-e.
We remove them for now because we don't use them but rely on the
gitian builds for Linux. We might readd adapted versions if we do
native Linux packaging in the future.

Also remove references to the removed files.

Fixes #648.

Signed-off-by: Cornelius Schumacher <cornelius@thirdhash.com>