Include coinbase in undo data. (#654)
This includes the coinbase in undo data and expands transaction checks to it.

In order to activate PoS and have the new coinbase use stake we need to include the coinbase in undo data for properly doing reorgs/connecting/disconnecting blocks. Bitcoin has a lot of checks which simply do "when coinbase then ignore" when it comes to undo data. This removes these and tries to have this separated out from #636 . This should not be merged before #636 is not done and stuff is maybe backported to this branch. Or we close this, unsure yet. But feel free to have a look.

The beatuy in this branch is that it's a strict refactoring and all functional tests and unit tests are green.

Signed-off-by: Julian Fleischer <julian@thirdhash.com>