Sort stakeable coins by amount, biggest first (#752)
As discussed with @Ruteri and @Gnappuraz .

This is useful in `pos-new-coinbase` as it is otherwise very hardly predictable which coins will be used when splitting stake, as is done currently to multiply number of stakeable coins. An example: When building 200 blocks (as is done in `rpc_blockchain.py`) which continuously produce rewards and stake from the UTXO set, after reward become mature, the `generate` call will either spend reward utxos or other utxos. This is basically not predictable as it simply uses (effectively) the first coin returned by stakeable coins - and currently the order depends on the order of transactions in `mapWalletTxs` which is essentially arbitrary as it is ordered by transaction id (which is as arbitrary as the order in a hashmap). It then becomes unpredictable which transactions will be spent completely (i.e. consumed completely, and thus do not show up in the utxo-transaction set (that is to say transactions that still harbor utxos)) and which ones are not.

Also a general improvement: The proposer will use coins in order of the amount, which should have higher chances of being eligible anyway.

Signed-off-by: Julian Fleischer <julian@thirdhash.com>