Make non-regular transactions standard (#755)
unit-e features different types of transactions. bitcoin only has one type of transaction and has a notion of whether it is "standard" or not. Unfortunately the default type of transaction is named "TxType::STANDARD" in unit-e which leads to confusion: There are standard standard transactions and non-standard standard transactions.

Thus this PR renames `TxType::STANDARD` to `TxType::REGULAR`.

It fixes #128 by considering all non-regular transactions to be standard. The notion of `IsStandard` really only applies to regular transactions and we don't want non-regular transactions not to be relayed by nodes. This does not occur in regtest or testnet as they do not require transactions to be standard.

Signed-off-by: Julian Fleischer <julian@thirdhash.com>