Remove Qt wallet references
Remove remaining references in docs and other files to the Qt
wallet. This includes removing the icons used by the wallet and
the script to optimize the icons.

Also remove the macdeploy references and related files as we
don't build the DMG anymore but use the same kind of tarball as
we use on Linux.

There still are some references in the translation handling but
these are unadapted documents anyway right now and the references
will go away once the translation process is adapted to unit-e.

Fixes #767.

Signed-off-by: Cornelius Schumacher <cornelius@thirdhash.com>