Make flaky test wallet_encryption.py a bit less flaky
The test used to fail sometimes, since it checked that the wallet
auto-locks itself after a timeout by sleeping for that amount of time
and then running an RPC call. The problem is that, when two processes
(the test and the daemon) sleep for the exact same amount of time, it's a
toss-up which one of them gets woken up first.

This commit increases the sleep time for the test by 0.1 seconds, which
should be enought of a safety margin for the check to pass always.

Signed-off-by: Mihai Ciumeica <mihai@thirdhash.com>