Correct kernel protocol (#774)
This corrects the kernel protocol implemented in unit-e. The kernel protocol was adapted from particl which in turn took it from [BlackCoin](https://blackcoin.org/blackcoin-pos-protocol-v2-whitepaper.pdf).

A simple description of PoS v3 can be found at http://earlz.net/view/2017/07/27/1904/the-missing-explanation-of-proof-of-stake-version

```
while(true) {
    foreach(utxo in wallet) {
        blockTime = currentTime - currentTime % 16
        posDifficulty = difficulty * utxo.value
        hash = hash(previousStakeModifier << utxo.time << utxo.hash << utxo.n << blockTime)
        if(hash < posDifficulty) {
            done
        }
    } 
    wait 16s -- wait 16 seconds, until the block time can be changed
}
```

The actual change is that currently in unit-e we take the time of the previous block (the block previous to the one being proposed, that is to say the current tip at the time of proposing), whereas PoS v3 uses the blocktime of the block which contains the stake (as can be seen in above pseudo code).

The wrong implementation in unit-e resulted from a misreading of particl code which mixes both in the description of the kernel protocol as well as in variable names `txPrev` and `prevout`. Also an earlier version of PoS (PoS v2) actually uses also the time of the previous block, but in the stake modifier. This is declared obsolete in PoS v3.

The calculation of the stake modifier has been corrected also: [particl code reads](https://github.com/particl/particl-core/blob/0.16/src/pos/kernel.cpp#L30):

```C++
uint256 ComputeStakeModifierV2(const CBlockIndex *pindexPrev, const uint256 &kernel)
{
    if (!pindexPrev)
        return uint256();  // genesis block's modifier is 0

    CDataStream ss(SER_GETHASH, 0);
    ss << kernel << pindexPrev->bnStakeModifier;
    return Hash(ss.begin(), ss.end());
}
```

But the variable named `kernel` here is not actually the kernel hash which is used to check whether a proposer was eligible to propose but the hash of the staking outpoint (i.e. the transaction that contains the referenced staking output).

The changes are in essence:

- [Use stake block time instead of previous' block time to compute kernel](https://github.com/dtr-org/unit-e/pull/774/files#diff-82f2ca5c047d98d2c3c917e18653ac0fR184)
- [Use stake transaction hash to compute stake modifier instead of kernel hash](https://github.com/dtr-org/unit-e/pull/774/files#diff-82f2ca5c047d98d2c3c917e18653ac0fR167)

As this requires to drag through different values `staking::Coin` was refactored to hold `CBlockIndex *` of the block containing the stake. At this occasion also the `depth` was changed to hold a `height`, which fixes #578 .

Signed-off-by: Julian Fleischer <julian@thirdhash.com>